fix: consolidate durable dispatch/artifact servertests for pkg-file-count limit
Merge dispatch, artifact, and parity API servertests into one file so
pkg/api/servertests stays within the 15-file CI cap.
